KP-4870: Added integration tests for ESENT

+++ src/Tests/WB.Tests.CleanIntegration/EsentTests/with_esent_store.cs
@@ -15,7 +15,10 @@ namespace WB.Tests.Integration.EsentTests
         Cleanup things = () =>
         {
             storage.Dispose();
-            PersistentDictionaryFile.DeleteFiles("TempStore");
+            if (PersistentDictionaryFile.Exists("TempStore"))
+            {
+                PersistentDictionaryFile.DeleteFiles("TempStore");
+            }
         };
 
         protected static EsentKeyValueStorage<T> storage;
